refactor: 체스 게임 규칙에 따라 작성된  로직은 ChessGame에서 관리
