+++ src/main/java/softeer2nd/chess/Board.java
@@ -130,28 +130,6 @@ public class Board {
         }
     }
 
-    public void move(String position, Piece piece) {
-        String[] px_py = position.split("");
-
-        int pos_x = px_py[0].charAt(0) - 'a';
-        int pos_y = 8 - Integer.parseInt(px_py[1]);
-
-        board.get(pos_y).set(pos_x, piece);
-    }
-
-    public void move(String sourcePosition, String targetPosition) {
-        String[] px_py_source = sourcePosition.split("");
-        String[] px_py_target = targetPosition.split("");
-
-        int pos_x_source = px_py_source[0].charAt(0) - 'a';
-        int pos_y_source = 8 - Integer.parseInt(px_py_source[1]);
-
-        int pos_x_target = px_py_target[0].charAt(0) - 'a';
-        int pos_y_target = 8 - Integer.parseInt(px_py_target[1]);
-
-        board.get(pos_y_target).set(pos_x_target, findPiece(sourcePosition));
-        board.get(pos_y_source).set(pos_x_source, Piece.createBlank());
-    }
 
     public double calculatePoint(Color color) {
         return board.stream()
